fake: Split fake transport and peer files
